Fix sphinx warnings and refactor code.
For autodoc docstrings to be linkable, they must be listed
in the APIdoc somewhere, therefore, they should be
public functions.

Therefore, I had to refactor the code base to
make some functions public (that arguably should
have been public already).

Now, there is a public "styling" module under plotting
and the `_utils` module is now public as `scirpy.util`.
