+++ .github/workflows/docs.yml
@@ -8,7 +8,6 @@ on:
   release:
     types: [created]
 
-
 jobs:
   docs:
     runs-on: ubuntu-latest
@@ -46,7 +45,7 @@ jobs:
           FLIT_ROOT_INSTALL=1 flit install --deps develop
       - name: run sphinx
         run: |
-          cd docs && make html   
+          cd docs && make html SPHINXOPTS="-W --keep-going"
       - name: Deploy to gh-pages
         uses: JamesIves/github-pages-deploy-action@releases/v3
         with:
