+++ scirpy/_plotting/_spectratype.py
@@ -3,7 +3,7 @@ import numpy as np
 from .._compat import Literal
 from anndata import AnnData
 from .. import tl
-from . import _base as base
+from . import base
 from typing import Union, List, Collection, Callable
