+++ scirpy/_preprocessing/_tcr_dist.py
@@ -1,5 +1,5 @@
 import parasail
-from .._util._multiprocessing import EnhancedPool as Pool
+from ..util._multiprocessing import EnhancedPool as Pool
 import itertools
 from anndata import AnnData
 from typing import Union, Collection, List, Tuple, Dict, Callable
@@ -7,7 +7,7 @@ from .._compat import Literal
 import numpy as np
 from scanpy import logging
 import numpy.testing as npt
-from .._util import _is_na, _is_symmetric, _reduce_nonzero
+from ..util import _is_na, _is_symmetric, _reduce_nonzero
 import abc
 from Levenshtein import distance as levenshtein_dist
 import scipy.spatial
