+++ scirpy/_tools/_group_abundance.py
@@ -2,7 +2,7 @@ from anndata import AnnData
 from typing import Union
 import numpy as np
 import pandas as pd
-from .._util import _is_na, _normalize_counts
+from ..util import _is_na, _normalize_counts
 from typing import Sequence
 from .._compat import Literal
