+++ tests/test_io.py
@@ -1,5 +1,5 @@
 from scirpy import read_10x_vdj, read_tracer
-from scirpy._util import _is_na, _is_false
+from scirpy.util import _is_na, _is_false
 import numpy as np
 import pytest
