+++ tests/test_tools.py
@@ -6,7 +6,7 @@ import pytest
 import numpy.testing as npt
 import pandas.testing as pdt
 import numpy as np
-from scirpy._util import _get_from_uns
+from scirpy.util import _get_from_uns
 from .fixtures import adata_clonotype, adata_tra, adata_vdj, adata_diversity
